Fix lint issues and update config
