+++ my-react-app/eslint.config.js
@@ -30,4 +30,10 @@ export default [
       ],
     },
   },
+  {
+    files: ['**/*.test.js'],
+    languageOptions: {
+      globals: { ...globals.jest },
+    },
+  },
 ]
