+++ src/components/Home/Spotlight/decorations/index.jsx
@@ -1,3 +1,3 @@
 // src/components/Home/Spotlight/decorations/index.js
-export { default as Stickers, SeasonalStickers } from './Stickers';
-export { default as PostItNotes, postItMessages } from './PostItNotes';
+export { default as Stickers } from './Stickers';
+export { default as PostItNotes } from './PostItNotes';
